Extract notification parameters to dedicated modules #5
